Modify Servo Driver and Constants
Still does not work, but after looking up servo motors, it turns out servo drivers rely on PWM. I tried modifying the constants to use a different range for pwm and angle. Also, I fixed the servo driver so that it properly parsed the input to an int (previously it would just throw an error and exit :P). It properly parses the input, but the updated duty cycle does not change the state of the motor at all. Also added more documentation to servo driver.