Add Separate File to Display Heatmap
As it stands, if we end up with manual control, we will probably simply copy the generated pickled file from the rpi to the laptop, so this program is a quick way to open the generated file. Perhaps eventually we can just open another socket and use the rpi to send the data directly from the consumer thread.